feat: adiciona conversao dinamica de moedas

+++ src/main/java/com/pet/model/ResultadoConversaoDto.java
@@ -3,11 +3,11 @@ package com.pet.model;
 import java.math.BigDecimal;
 
 public record ResultadoConversaoDto(
-			Double valorOriginal,
+			BigDecimal valorOriginal,
 		    String moedaOrigem,
 		    String moedaDestino,
-		    Double taxaCambio,
-		    Double valorConvertido
+		    BigDecimal taxaCambio,
+		    BigDecimal valorConvertido
 		) {
 
 }
